requirements.txt
python-dotenv
folium
streamlit-folium
geopy
google-generativeai